Remove distinction between synchronous plugins and asynchronous ones
This commit is a major refactor of the code. It removes the async keyword
such that all plugins are now part of the same array, `_plugins`.
Asynchronicity is now indicated with the `defer` keyword. This has the
benefit that we can now use the modifiers that `zsh-defer` gives us.

`ignore` and `defer` are mutually exclusive. That is because we only call
`zsh-defer` on the sourcing command itself, which therefore makes no
sense when the `ignore` flag is set.

Another improvement is that the case statement now has been greatly
simplified as `preload`, `postload` and `postinstall` now share the same
code.